+++ lib/src/presentaion/widgets/sheets/new_quote_sheet.dart
@@ -16,12 +16,12 @@ class NewQuoteSheet extends StatefulWidget {
   final List<String>? servicesList;
   final String? date;
   final String? time;
-  final double defaultValue;
+  final int defaultValue;
   final NewQuoteSheetSheetVariant variant;
   final int? workOrderId;
   //final MerchantHomeScreenCubit? cubit;
   final VoidCallback onCancel;
-  final Function(double) onSubmit;
+  final Function(int) onSubmit;
 
   const NewQuoteSheet({
     super.key,
@@ -45,7 +45,7 @@ class NewQuoteSheet extends StatefulWidget {
 }
 
 class _NewQuoteSheetState extends State<NewQuoteSheet> {
-  double _defaultValue = 0;
+  int _defaultValue = 0;
   late TextEditingController _amountController;
   final int minAmount = 100;
   final int maxAmount = 1000; //might be needed from backend
@@ -172,7 +172,7 @@ class _NewQuoteSheetState extends State<NewQuoteSheet> {
                         onChanged: (value) {
                           setState(() {
                             _defaultValue =
-                                double.tryParse(value) ?? _defaultValue;
+                                int.tryParse(value) ?? _defaultValue;
                             _errorMessage = null;
                           });
                         },
@@ -258,7 +258,7 @@ class _NewQuoteSheetState extends State<NewQuoteSheet> {
                         onCancelTap: () {
                           // widget.cubit!.cancelWorkOrder(widget.workOrderId ?? -1, getIt<UserEntity>().id);
 
-                          widget.onSubmit(0.0);
+                          widget.onSubmit(0);
                           context.pop();
 
                           //context.popUntil(RouteConstants.merchantHomeScreenRoute);
